Update navigation framework
