+++ components/DocsNav.tsx
@@ -4,16 +4,7 @@ import { Text, Box } from '@modulz/design-system';
 
 export function NavHeading({ children, ...props }: { children: React.ReactNode }) {
   return (
-    <Text
-      as="h4"
-      size="3"
-      css={{
-        fontWeight: 500,
-        px: '$5',
-        py: '$2',
-      }}
-      {...props}
-    >
+    <Text as="h4" size="3" css={{ fontWeight: 500, px: '$3', py: '$2' }} {...props}>
       {children}
     </Text>
   );
@@ -44,8 +35,9 @@ export function NavItem({ children, active, disabled, href, ...props }: NavItemP
           textDecoration: 'none',
           color: disabled ? '$gray10' : '$hiContrast',
           py: '$2',
-          px: '$5',
+          px: '$3',
           backgroundColor: active ? '$violet5' : 'transparent',
+          borderRadius: 9999,
           userSelect: 'none',
           minHeight: '$6',
           transition: 'background-color 50ms linear',
