--- data/design-system/components/button/0.0.1.mdx
@@ -1,6 +1,0 @@
----
-metaTitle: Button
-name: button
----
-
-Coming soon.
