--- data/design-system/overview/introduction.mdx
@@ -1,11 +1,0 @@
----
-metaTitle: Introduction
-metaDescription: An open-source design system for building high-quality, accessible design systems and web apps.
----
-
-# Introduction
-
-<Description>
-  An open-source design system for building high-quality, accessible design
-  systems and web apps.
-</Description>
