--- pages/docs/design-system/overview/[slug].tsx
@@ -1,76 +1,0 @@
-import React from 'react';
-import { getMDXComponent } from 'mdx-bundler/client';
-import { Box } from '@modulz/design-system';
-import { TitleAndMetaTags } from '@components/TitleAndMetaTags';
-import { MDXProvider, components } from '@components/MDXComponents';
-import { getAllFrontmatter, getMdxBySlug } from '@lib/mdx';
-import { RemoveScroll } from 'react-remove-scroll';
-import { QuickNav } from '@components/QuickNav';
-
-import type { Frontmatter } from 'types/frontmatter';
-
-type Doc = {
-  frontmatter: Frontmatter;
-  code: any;
-};
-
-export default function DesignSystemOverviewDoc({ frontmatter, code }: Doc) {
-  const Component = React.useMemo(() => getMDXComponent(code), [code]);
-
-  return (
-    <>
-      <TitleAndMetaTags
-        title={`${frontmatter.metaTitle} – Radix UI`}
-        description={frontmatter.metaDescription}
-        image={frontmatter.metaImage}
-      />
-
-      <MDXProvider frontmatter={frontmatter}>
-        <Component components={components as any} />
-      </MDXProvider>
-
-      <Box
-        as="aside"
-        // Components that hide the scrollbar (like Dialog) add padding to
-        // account for the scrollbar gap to avoid layout jank. This does not
-        // work for position: fixed elements. Since we use react-remove-scroll
-        // under the hood for those primitives, we can add this helper class
-        // provided by that lib to deal with that for the QuickNav.
-        // https://github.com/radix-ui/website/issues/64
-        // https://github.com/theKashey/react-remove-scroll#positionfixed-elements
-        className={RemoveScroll.classNames.zeroRight}
-        css={{
-          display: 'none',
-          '@media (min-width: 1440px)': {
-            display: 'block',
-            width: 250,
-            flexShrink: 0,
-            zIndex: 1,
-            position: 'fixed',
-            top: '$sizes$8',
-            right: 0,
-            bottom: 0,
-          },
-        }}
-      >
-        <QuickNav key={frontmatter.slug} />
-      </Box>
-    </>
-  );
-}
-
-export async function getStaticPaths() {
-  const frontmatters = getAllFrontmatter('design-system/overview');
-
-  return {
-    paths: frontmatters.map((frontmatter) => ({
-      params: { slug: frontmatter.slug.replace('design-system/overview/', '') },
-    })),
-    fallback: false,
-  };
-}
-
-export async function getStaticProps(context) {
-  const { frontmatter, code } = await getMdxBySlug('design-system/overview/', context.params.slug);
-  return { props: { frontmatter, code } };
-}
